requirements.txt
keyboard
pyautogui
requests
pillow
opencv-python